requirements.txt
pulsar-client
flask
requests